refactor: extract health, dispatch, and chat modules from daemon/mod.rs [Midtown #7]
Extract three domain-specific modules from the monolithic mod.rs:

- health.rs (~860 lines): Health checks — idle shutdown, nudge handling,
  stuck detection, zombie respawning, usage limits, reminders
- dispatch.rs (~770 lines): Task coordination — orphan recovery, duplicate
  worker detection, pending task spawning, worktree cleanup
- chat.rs (~230 lines): Chat monitor — @mention routing, @all broadcasts

Updates call sites in events.rs, rpc.rs, and mod.rs. Moves associated
tests alongside their functions. mod.rs shrinks from ~5,850 to ~4,060
lines (~1,800 line reduction).

No behavioral changes. All 465 unit tests pass.
